заменила vechmach классы на классы из Math

+++ src/test/java/com/cgvsu/vectormath/matrix/Matrix3x3Test.java
@@ -1,20 +1,20 @@
 package com.cgvsu.vectormath.matrix;
 
 import com.cgvsu.vectormath.vector.Vector3D;
-import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.*;
+
+import org.testng.annotations.Test;
 
 public class Matrix3x3Test {
 
     @Test
     public void testAddition() {
-        double[][] data1 = {
+        float[][] data1 = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
         };
-        double[][] data2 = {
+        float[][] data2 = {
                 {9, 8, 7},
                 {6, 5, 4},
                 {3, 2, 1}
@@ -24,7 +24,7 @@ public class Matrix3x3Test {
 
         Matrix3x3 result = m1.add(m2);
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {10, 10, 10},
                 {10, 10, 10},
                 {10, 10, 10}
@@ -39,12 +39,12 @@ public class Matrix3x3Test {
 
     @Test
     public void testSubtraction() {
-        double[][] data1 = {
+        float[][] data1 = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
         };
-        double[][] data2 = {
+        float[][] data2 = {
                 {9, 8, 7},
                 {6, 5, 4},
                 {3, 2, 1}
@@ -54,7 +54,7 @@ public class Matrix3x3Test {
 
         Matrix3x3 result = m1.subtract(m2);
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {-8, -6, -4},
                 {-2, 0, 2},
                 {4, 6, 8}
@@ -70,7 +70,7 @@ public class Matrix3x3Test {
 
     @Test
     public void testVectorMultiplication() {
-        double[][] data = {
+        float[][] data = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
@@ -89,12 +89,12 @@ public class Matrix3x3Test {
 
     @Test
     public void testMatrixMultiplication() {
-        double[][] data1 = {
+        float[][] data1 = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
         };
-        double[][] data2 = {
+        float[][] data2 = {
                 {9, 8, 7},
                 {6, 5, 4},
                 {3, 2, 1}
@@ -104,7 +104,7 @@ public class Matrix3x3Test {
 
         Matrix3x3 result = m1.multiply(m2);
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {30, 24, 18},
                 {84, 69, 54},
                 {138, 114, 90}
@@ -120,7 +120,7 @@ public class Matrix3x3Test {
 
     @Test
     public void testTranspose() {
-        double[][] data = {
+        float[][] data = {
                 {1, 2, 3},
                 {4, 5, 6},
                 {7, 8, 9}
@@ -129,7 +129,7 @@ public class Matrix3x3Test {
 
         Matrix3x3 result = matrix.transpose();
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {1, 4, 7},
                 {2, 5, 8},
                 {3, 6, 9}
@@ -147,13 +147,13 @@ public class Matrix3x3Test {
     public void testIdentityMatrix() {
         Matrix3x3 identity = Matrix3x3.identity();
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {1, 0, 0},
                 {0, 1, 0},
                 {0, 0, 1}
         };
 
-        double[][] actualData = identity.getMatrix();
+        float[][] actualData = identity.getMatrix();
 
         assertArrayEquals(expectedData, actualData);
     }
@@ -162,26 +162,26 @@ public class Matrix3x3Test {
     public void testZeroMatrix() {
         Matrix3x3 zero = Matrix3x3.zero();
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {0, 0, 0},
                 {0, 0, 0},
                 {0, 0, 0}
         };
 
-        double[][] actualData = zero.getMatrix();
+        float[][] actualData = zero.getMatrix();
 
         assertArrayEquals(expectedData, actualData);
     }
     @Test
     public void testDetermination() {
-        double[][] data = {
+        float[][] data = {
                 {10, 34, 5},
                 {28, 12, 93},
                 {7, 48, 200}
         };
         Matrix3x3 matrix = new Matrix3x3(data);
 
-        double result = matrix.determinate();
+        float result = matrix.determinate();
         assertEquals(-182606, result, 0.001);
     }
 }
