+++ src/test/java/com/cgvsu/vectormath/matrix/Matrix4x4Test.java
@@ -1,21 +1,22 @@
 package com.cgvsu.vectormath.matrix;
 
 import com.cgvsu.vectormath.vector.Vector4D;
-import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.*;
+
+import org.testng.annotations.Test;
+
 
 public class Matrix4x4Test {
 
     @Test
     public void testAddition() {
-        double[][] data1 = {
+        float[][] data1 = {
                 {1, 2, 3, 4},
                 {5, 6, 7, 8},
                 {9, 10, 11, 12},
                 {13, 14, 15, 16}
         };
-        double[][] data2 = {
+        float[][] data2 = {
                 {16, 15, 14, 13},
                 {12, 11, 10, 9},
                 {8, 7, 6, 5},
@@ -26,7 +27,7 @@ public class Matrix4x4Test {
 
         Matrix4x4 result = m1.add(m2);
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {17, 17, 17, 17},
                 {17, 17, 17, 17},
                 {17, 17, 17, 17},
@@ -37,13 +38,13 @@ public class Matrix4x4Test {
 
     @Test
     public void testSubtraction() {
-        double[][] data1 = {
+        float[][] data1 = {
                 {1, 2, 3, 4},
                 {5, 6, 7, 8},
                 {9, 10, 11, 12},
                 {13, 14, 15, 16}
         };
-        double[][] data2 = {
+        float[][] data2 = {
                 {16, 15, 14, 13},
                 {12, 11, 10, 9},
                 {8, 7, 6, 5},
@@ -54,7 +55,7 @@ public class Matrix4x4Test {
 
         Matrix4x4 result = m1.subtract(m2);
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {-15, -13, -11, -9},
                 {-7, -5, -3, -1},
                 {1, 3, 5, 7},
@@ -65,7 +66,7 @@ public class Matrix4x4Test {
 
     @Test
     public void testMatrixVectorMultiplication() {
-        double[][] matrixData = {
+        float[][] matrixData = {
                 {2, 0, 0, 0},
                 {0, 2, 0, 0},
                 {0, 0, 2, 0},
@@ -73,7 +74,7 @@ public class Matrix4x4Test {
         };
         Matrix4x4 matrix = new Matrix4x4(matrixData);
 
-        double[] vectorData = {1, 2, 3, 4};
+        float[] vectorData = {1, 2, 3, 4};
         Vector4D vector = new Vector4D(vectorData[0], vectorData[1],vectorData[2], vectorData[3]);
 
         Vector4D result = matrix.multiply(vector);
@@ -86,7 +87,7 @@ public class Matrix4x4Test {
 
     @Test
     public void testMatrixMatrixMultiplication() {
-        double[][] matrixData1 = {
+        float[][] matrixData1 = {
                 {1, 2, 3, 4},
                 {5, 6, 7, 8},
                 {9, 10, 11, 12},
@@ -94,7 +95,7 @@ public class Matrix4x4Test {
         };
         Matrix4x4 matrix1 = new Matrix4x4(matrixData1);
 
-        double[][] matrixData2 = {
+        float[][] matrixData2 = {
                 {16, 15, 14, 13},
                 {12, 11, 10, 9},
                 {8, 7, 6, 5},
@@ -104,7 +105,7 @@ public class Matrix4x4Test {
 
         Matrix4x4 result = matrix1.multiply(matrix2);
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {80, 70, 60, 50},
                 {240, 214, 188, 162},
                 {400, 358, 316, 274},
@@ -115,7 +116,7 @@ public class Matrix4x4Test {
 
     @Test
     public void testTranspose() {
-        double[][] matrixData = {
+        float[][] matrixData = {
                 {1, 2, 3, 4},
                 {5, 6, 7, 8},
                 {9, 10, 11, 12},
@@ -125,7 +126,7 @@ public class Matrix4x4Test {
 
         Matrix4x4 transposed = matrix.transpose();
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {1, 5, 9, 13},
                 {2, 6, 10, 14},
                 {3, 7, 11, 15},
@@ -137,9 +138,9 @@ public class Matrix4x4Test {
     @Test
     public void testIdentityMatrix() {
         Matrix4x4 identity = Matrix4x4.identity();
-        double[][] data = identity.getMatrix();
+        float[][] data = identity.getMatrix();
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {1, 0, 0, 0},
                 {0, 1, 0, 0},
                 {0, 0, 1, 0},
@@ -152,9 +153,9 @@ public class Matrix4x4Test {
     @Test
     public void testZeroMatrix() {
         Matrix4x4 zero = Matrix4x4.zero();
-        double[][] data = zero.getMatrix();
+        float[][] data = zero.getMatrix();
 
-        double[][] expectedData = {
+        float[][] expectedData = {
                 {0, 0, 0, 0},
                 {0, 0, 0, 0},
                 {0, 0, 0, 0},
@@ -165,7 +166,7 @@ public class Matrix4x4Test {
     }
     @Test
     public void testDetermination() {
-        double[][] data = {
+        float[][] data = {
                 {10, 34, 5, 45},
                 {28, 12, 93, 3},
                 {7, 48, 200, 10},
@@ -173,12 +174,12 @@ public class Matrix4x4Test {
         };
         Matrix4x4 matrix = new Matrix4x4(data);
 
-        double result = matrix.determinate();
+        float result = matrix.determinate();
         assertEquals(308450, result, 0.001);
     }
     @Test
     public void testDetermination1() {
-        double[][] data = {
+        float[][] data = {
                 {1, 7, 1, 8},
                 {4, 6, 11, 8},
                 {1, 3, 1, 8},
@@ -186,7 +187,7 @@ public class Matrix4x4Test {
         };
         Matrix4x4 matrix = new Matrix4x4(data);
 
-        double result = matrix.determinate();
+        float result = matrix.determinate();
         assertEquals(252, result, 0.001);
     }
 }
